refactoring, duplicate user blocks
