+++ focusdriving/src/main/java/com/cmpt/focusdriving/models/Student/StudentRepository.java
@@ -1,4 +1,4 @@
-package com.cmpt.focusdriving.models;
+package com.cmpt.focusdriving.models.Student;
 
 import java.util.List;
